# api-spec/tsp-output/@typespec/openapi3/openapi.api.1.0.yaml
openapi: 3.0.0
info:
  title: Microservice API
  version: '1.0'
...
paths:
...
  /api/{version}/login:
    post:
      operationId: LoginController_login
      description: user login
      parameters:
        - $ref: '#/components/parameters/Utils.Versioning.VersionParams'
      responses:
        '200':
          description: login successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login.LoginOutput'
        '400':
          description: When login input is invalid
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Login.LoginValidationError'
                  - $ref: '#/components/schemas/Login.LoginWrongPasswordError'
        '404':
          description: When user not found
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Login.LoginUserNotFoundError'
                  - $ref: '#/components/schemas/Login.LoginRoleNotFoundError'
      tags:
        - Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login.LoginInput'
  /api/{version}/login/google:
    get:
      operationId: LoginController_loginGoogle
      description: Google OAuth login redirect
      parameters:
        - $ref: '#/components/parameters/Utils.Versioning.VersionParams'
      responses:
        '204':
          description: Redirects to Google OAuth
      tags:
        - Login
  /api/{version}/login/google/callback:
    get:
      operationId: LoginController_loginGoogleCallback
      description: Google OAuth callback
      parameters:
        - $ref: '#/components/parameters/Utils.Versioning.VersionParams'
      responses:
        '204':
          description: Handles Google OAuth callback
      tags:
        - Login
...
  /api/{version}/refresh:
    post:
      operationId: LoginController_refresh
      description: get user refresh token
      parameters:
        - $ref: '#/components/parameters/Utils.Versioning.VersionParams'
      responses:
        '200':
          description: login refresh token successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login.RefreshTokenOutput'
      tags:
        - Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login.RefreshTokenInput'
...
components:
  parameters:
...
    Utils.Versioning.VersionParams:
      name: version
      in: path
      required: true
      description: route version
      schema:
        $ref: '#/components/schemas/Utils.Versioning.ApiVersion'
  schemas:
...
    Login.LoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          default: admin@admin.com
        password:
          type: string
          default: admin
      description: login input
    Login.LoginOutput:
      type: object
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      description: login ouput
    Login.LoginRoleNotFoundError:
      type: object
      allOf:
        - type: object
          required:
            - error
          properties:
            error:
              type: object
              required:
                - code
                - traceid
                - context
                - message
                - timestamp
                - path
              properties:
                code:
                  type: number
                  enum:
                    - 404
                  description: status code
                traceid:
                  type: string
                  description: request traceid
                context:
                  type: string
                  description: class that error occur
                message:
                  type: array
                  items:
                    type: string
                    enum:
                      - roleNotFound
                  description: error message
                timestamp:
                  type: string
                  description: timestamp that error occur
                path:
                  type: string
                  description: path error
              description: default error model
          description: When resource not found.
      description: When user role not found
    Login.LoginUserNotFoundError:
      type: object
      allOf:
        - type: object
          required:
            - error
          properties:
            error:
              type: object
              required:
                - code
                - traceid
                - context
                - message
                - timestamp
                - path
              properties:
                code:
                  type: number
                  enum:
                    - 404
                  description: status code
                traceid:
                  type: string
                  description: request traceid
                context:
                  type: string
                  description: class that error occur
                message:
                  type: array
                  items:
                    type: string
                    enum:
                      - userNotFound
                  description: error message
                timestamp:
                  type: string
                  description: timestamp that error occur
                path:
                  type: string
                  description: path error
              description: default error model
          description: When resource not found.
      description: When user not found
    Login.LoginValidationError:
      type: object
      allOf:
        - type: object
          required:
            - error
          properties:
            error:
              type: object
              required:
                - code
                - traceid
                - context
                - message
                - timestamp
                - path
              properties:
                code:
                  type: number
                  enum:
                    - 400
                  description: status code
                traceid:
                  type: string
                  description: request traceid
                context:
                  type: string
                  description: class that error occur
                message:
                  type: array
                  items:
                    type: string
                    enum:
                      - 'email: Required'
                      - 'password: Required'
                  description: error message
                timestamp:
                  type: string
                  description: timestamp that error occur
                path:
                  type: string
                  description: path error
              description: default error model
          description: When input is invalid.
      description: When login input is invalid
    Login.LoginWrongPasswordError:
      type: object
      allOf:
        - type: object
          required:
            - error
          properties:
            error:
              type: object
              required:
                - code
                - traceid
                - context
                - message
                - timestamp
                - path
              properties:
                code:
                  type: number
                  enum:
                    - 400
                  description: status code
                traceid:
                  type: string
                  description: request traceid
                context:
                  type: string
                  description: class that error occur
                message:
                  type: array
                  items:
                    type: string
                    enum:
                      - incorrectPassword
                  description: error message
                timestamp:
                  type: string
                  description: timestamp that error occur
                path:
                  type: string
                  description: path error
              description: default error model
          description: When input is invalid.
      description: When user password is wrong
    Login.RefreshTokenInput:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
      description: login refresh token input
    Login.RefreshTokenOutput:
      type: object
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      description: login refresh token output
...
    Utils.Versioning.ApiVersion:
      type: string
      enum:
        - v1
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer
servers:
  - url: http://localhost:5000/
    variables: {}